requirements.txt
# Recommended runtime: Python 3.12.x (PyMuPDF may fail on 3.14 on Windows).
fastapi==0.115.2
uvicorn==0.30.6
python-multipart==0.0.9
pydantic==2.9.2
pydantic-settings==2.6.1
python-docx==1.1.2
PyMuPDF==1.24.10
Pillow==10.4.0
pytesseract==0.3.13
openai>=1.0.0
